+++ src/NFluent/Messages/FluentMessage.cs
@@ -154,7 +154,7 @@ namespace NFluent.Extensibility
         /// <returns>
         /// A <see cref="FluentMessage"/> to continue build the message.
         /// </returns>
-        public MessageBlock On(object test, int index = 0)
+        public MessageBlock On(object test, long index = 0)
         {
             this.checkedBlock = new MessageBlock(this, test, this.checkedLabel, index);
             if (this.referenceType == null)
@@ -218,7 +218,7 @@ namespace NFluent.Extensibility
         /// <returns>
         /// The created MessageBlock.
         /// </returns>
-        public MessageBlock ExpectedValues(object expectedValues, int index = 0)
+        public MessageBlock ExpectedValues(object expectedValues, long index = 0)
         {
             this.expectedLabel = GenericLabelBlock.BuildExpectedBlock(new EntityNamer { EntityName = "value(s)" });
             this.expectedBlock = new MessageBlock(this, expectedValues, this.expectedLabel, index);
